Extract logic from Job to Service object
