+++ app/services/streams/twitch_webhook/register.rb
@@ -1,0 +1,41 @@
+module Streams
+  module TwitchWebhook
+    class Register
+      WEBHOOK_LEASE_SECONDS = 7.days.to_i
+
+      def initialize(user)
+        @user = user
+      end
+
+      def self.call(*args)
+        new(*args).call
+      end
+
+      def call
+        user_resp = client.get("users", login: user.twitch_username)
+        twitch_user_id = user_resp.body["data"].first["id"]
+
+        client.post(
+          "webhooks/hub",
+          "hub.callback" => twitch_stream_updates_url_for_user(user),
+          "hub.mode" => "subscribe",
+          "hub.lease_seconds" => WEBHOOK_LEASE_SECONDS,
+          "hub.topic" => "https://api.twitch.tv/helix/streams?user_id=#{twitch_user_id}",
+          "hub.secret" => ApplicationConfig["TWITCH_WEBHOOK_SECRET"],
+        )
+      end
+
+      private
+
+      attr_reader :user
+
+      def client
+        @client ||= Streams::TwitchCredentials.generate_client
+      end
+
+      def twitch_stream_updates_url_for_user(user)
+        Rails.application.routes.url_helpers.user_twitch_stream_updates_url(user_id: user.id, host: ApplicationConfig["APP_DOMAIN"])
+      end
+    end
+  end
+end
