+++ spec/jobs/streams/twitch_webhook_registration_job_spec.rb
@@ -3,55 +3,33 @@ require "rails_helper"
 RSpec.describe Streams::TwitchWebhookRegistrationJob, type: :job do
   let(:user) { create(:user, twitch_username: "test-username") }
 
-  let(:expected_headers) do
-    { "Authorization" => "Bearer FAKE_TWITCH_TOKEN" }
-  end
-  let(:expected_twitch_webhook_params) do
-    {
-      "hub.callback" => "http://#{ApplicationConfig['APP_DOMAIN']}/users/#{user.id}/twitch_stream_updates",
-      "hub.mode" => "subscribe",
-      "hub.lease_seconds" => 604_800,
-      "hub.topic" => "https://api.twitch.tv/helix/streams?user_id=654321",
-      "hub.secret" => ApplicationConfig["TWITCH_WEBHOOK_SECRET"]
-    }
-  end
-  let!(:twitch_webhook_registration_stubbed_route) do
-    stub_request(:post, "https://api.twitch.tv/helix/webhooks/hub").
-      with(body: expected_twitch_webhook_params, headers: expected_headers).
-      and_return(status: 204)
-  end
+  let(:service) { double }
 
-  let(:expected_twitch_user_params) { { login: "test-username" } }
-  let!(:twitch_user_stubbed_route) do
-    stub_request(:get, "https://api.twitch.tv/helix/users").
-      with(query: expected_twitch_user_params, headers: expected_headers).
-      and_return(body: { data: [{ id: 654_321 }] }.to_json)
+  before do
+    allow(service).to receive(:call)
   end
 
   context "when the user does NOT have a twitch username present" do
     let(:user) { create(:user) }
 
     it "noops" do
-      described_class.perform_now(user.id)
+      described_class.perform_now(user.id, service)
 
-      expect(twitch_webhook_registration_stubbed_route).not_to have_been_requested
-      expect(twitch_user_stubbed_route).not_to have_been_requested
+      expect(service).not_to have_received(:call)
     end
   end
 
   it "noops when the id passed does not belong to a user" do
-    described_class.perform_now(987_654_321)
+    described_class.perform_now(987_654_321, service)
 
-    expect(twitch_webhook_registration_stubbed_route).not_to have_been_requested
-    expect(twitch_user_stubbed_route).not_to have_been_requested
+    expect(service).not_to have_received(:call)
   end
 
   it "registers for webhooks" do
     allow(Streams::TwitchCredentials).to receive(:access_token).and_return("FAKE_TWITCH_TOKEN")
 
-    described_class.perform_now(user.id)
+    described_class.perform_now(user.id, service)
 
-    expect(twitch_webhook_registration_stubbed_route).to have_been_requested
-    expect(twitch_user_stubbed_route).to have_been_requested
+    expect(service).to have_received(:call).with(user)
   end
 end
